Declaring variables inside @implementation but without { } makes them extern and global, not ivars
This is really dangerous in case you were to allocate more than one VC of this type. They would both rely on the same storage for those variables and hilarity would ensue.

Also because they don't belong to objects of the class, ARC wouldn't release them when an object of the class is deallocated.


This is not the only issue I've found in the code, but I thought it was worth pointing out.